shift and test
