+++ vscode/src/editor/utils/editor-context.test.ts
@@ -1,5 +1,6 @@
 import { afterEach, describe, expect, it, vi } from 'vitest'
 import * as vscode from 'vscode'
+import { URI } from 'vscode-uri'
 
 import {
     type ContextItemFile,
@@ -7,10 +8,12 @@ import {
     ignores,
     testFileUri,
     uriBasename,
+    type ContextItem,
+    type Editor,
 } from '@sourcegraph/cody-shared'
 
 import { CHARS_PER_TOKEN } from '@sourcegraph/cody-shared/src/prompt/constants'
-import { filterLargeFiles, getFileContextFiles } from './editor-context'
+import { fillInContextItemContent, filterLargeFiles, getFileContextFiles } from './editor-context'
 
 vi.mock('lodash/throttle', () => ({ default: vi.fn(fn => fn) }))
 
@@ -177,5 +180,32 @@ describe('filterLargeFiles', () => {
         const filtered = await filterLargeFiles([largeTextFile])
 
         expect(filtered[0].title).toEqual('large-file')
+    })
+})
+
+describe('fillInContextItemContent', () => {
+    it('omits files that could not be read', async () => {
+        // Fixes https://github.com/sourcegraph/cody/issues/2390.
+        const mockEditor: Partial<Editor> = {
+            getTextEditorContentForFile(uri) {
+                if (uri.path === '/a.txt') {
+                    return Promise.resolve('a')
+                }
+                throw new Error('error')
+            },
+        }
+        const contextItems = await fillInContextItemContent(mockEditor as Editor, [
+            {
+                type: 'file',
+                uri: URI.parse('file:///a.txt'),
+            },
+            {
+                type: 'file',
+                uri: URI.parse('file:///error.txt'),
+            },
+        ])
+        expect(contextItems).toEqual<ContextItem[]>([
+            { type: 'file', uri: URI.parse('file:///a.txt'), content: 'a' },
+        ])
     })
 })
